fix(chat): render completion links in overlay
