+++ packages/ui/src/components/chat/MessageContent.tsx
@@ -73,6 +73,7 @@ import {
   sensitiveRequestTitleLabel,
   splitInlineCode,
 } from "./message-parser-helpers";
+import { SafeMarkdownText } from "./SafeMarkdownText";
 import { ThinkingBlock } from "./ThinkingBlock";
 import { useParsedSegments } from "./use-parsed-segments";
 import { ChatWidgetShell } from "./widgets/chat-widget-shell";
@@ -108,49 +109,8 @@ function isHttpsAuthorizationUrl(url: unknown): url is string {
   }
 }
 
-function safeChatLinkTarget(raw: string): string | null {
-  if (raw.startsWith("/") && !raw.startsWith("//")) return raw;
-  try {
-    const url = new URL(raw);
-    return url.protocol === "https:" || url.protocol === "http:" ? raw : null;
-  } catch {
-    return null;
-  }
-}
-
-function renderLinkedText(text: string, keyPrefix: string): ReactNode {
-  const linkPattern = /\[([^\]\n]+)\]\(([^)\s]+)\)/g;
-  const nodes: ReactNode[] = [];
-  let cursor = 0;
-  for (const match of text.matchAll(linkPattern)) {
-    const index = match.index ?? 0;
-    const [raw, label, rawTarget] = match;
-    if (index > cursor) nodes.push(text.slice(cursor, index));
-    const target = safeChatLinkTarget(rawTarget);
-    nodes.push(
-      target ? (
-        <a
-          key={`${keyPrefix}:link:${index}`}
-          href={target}
-          target="_blank"
-          rel="noreferrer"
-          className="text-accent underline underline-offset-2"
-        >
-          {label}
-        </a>
-      ) : (
-        raw
-      ),
-    );
-    cursor = index + raw.length;
-  }
-  if (cursor === 0) return text;
-  if (cursor < text.length) nodes.push(text.slice(cursor));
-  return nodes;
-}
-
 function renderInlineText(text: string): ReactNode {
-  if (!text.includes("`")) return renderLinkedText(text, "text");
+  if (!text.includes("`")) return <SafeMarkdownText text={text} />;
   const parts = splitInlineCode(text);
   if (parts.length === 1 && parts[0].kind === "text") return text;
   // Key by the part's character offset in the source run (not the array index)
@@ -168,7 +128,9 @@ function renderInlineText(text: string): ReactNode {
         data-testid="inline-code"
       />
     ) : (
-      <span key={key}>{renderLinkedText(part.text, key)}</span>
+      <span key={key}>
+        <SafeMarkdownText text={part.text} />
+      </span>
     );
   });
 }
